Remove print statements and add tests
